data_agent_viz/requirements.txt
google-cloud-aiplatform[adk,agent-engines]


